Refactor value vs value transpilers (#142)
