+++ src/Assertion/AbstractTwoValueComparisonTranspiler.php
@@ -1,0 +1,97 @@
+<?php declare(strict_types=1);
+
+namespace webignition\BasilTranspiler\Assertion;
+
+use webignition\BasilModel\Assertion\AssertionInterface;
+use webignition\BasilTranspiler\CallFactory\AssertionCallFactory;
+use webignition\BasilTranspiler\CallFactory\VariableAssignmentCallFactory;
+use webignition\BasilTranspiler\Model\Call\VariableAssignmentCall;
+use webignition\BasilTranspiler\Model\TranspilationResultInterface;
+use webignition\BasilTranspiler\Model\VariablePlaceholder;
+use webignition\BasilTranspiler\NonTranspilableModelException;
+use webignition\BasilTranspiler\TranspilerInterface;
+
+abstract class AbstractTwoValueComparisonTranspiler implements TranspilerInterface
+{
+    protected $assertionCallFactory;
+    private $variableAssignmentCallFactory;
+    private $assertableValueExaminer;
+
+    public function __construct(
+        AssertionCallFactory $assertionCallFactory,
+        VariableAssignmentCallFactory $variableAssignmentCallFactory,
+        AssertableValueExaminer $assertableValueExaminer
+    ) {
+        $this->assertionCallFactory = $assertionCallFactory;
+        $this->variableAssignmentCallFactory = $variableAssignmentCallFactory;
+        $this->assertableValueExaminer = $assertableValueExaminer;
+    }
+
+    abstract protected function getHandledComparisons(): array;
+    abstract protected function getAssertionCall(
+        string $comparison,
+        VariableAssignmentCall $examinedValue,
+        VariableAssignmentCall $expectedValue
+    ): TranspilationResultInterface;
+
+    public function handles(object $model): bool
+    {
+        if (!$model instanceof AssertionInterface) {
+            return false;
+        }
+
+        return in_array($model->getComparison(), $this->getHandledComparisons());
+    }
+
+    /**
+     * @param object $model
+     *
+     * @return TranspilationResultInterface
+     *
+     * @throws NonTranspilableModelException
+     */
+    public function transpile(object $model): TranspilationResultInterface
+    {
+        if (!$model instanceof AssertionInterface) {
+            throw new NonTranspilableModelException($model);
+        }
+
+        $isHandledComparison = in_array($model->getComparison(), $this->getHandledComparisons());
+
+        if (false === $isHandledComparison) {
+            throw new NonTranspilableModelException($model);
+        }
+
+        $examinedValue = $model->getExaminedValue();
+        if (null === $examinedValue || !$this->assertableValueExaminer->isAssertableExaminedValue($examinedValue)) {
+            throw new NonTranspilableModelException($model);
+        }
+
+        $expectedValue = $model->getExpectedValue();
+        if (null === $expectedValue || !$this->assertableValueExaminer->isAssertableExpectedValue($expectedValue)) {
+            throw new NonTranspilableModelException($model);
+        }
+
+        $examinedValuePlaceholder = new VariablePlaceholder('EXAMINED_VALUE');
+        $examinedValueAssignmentCall = $this->variableAssignmentCallFactory->createValueVariableAssignmentCall(
+            $examinedValue,
+            $examinedValuePlaceholder
+        );
+
+        $expectedValuePlaceholder = new VariablePlaceholder('EXPECTED_VALUE');
+        $expectedValueAssignmentCall = $this->variableAssignmentCallFactory->createValueVariableAssignmentCall(
+            $expectedValue,
+            $expectedValuePlaceholder
+        );
+
+        if (null === $expectedValueAssignmentCall || null === $examinedValueAssignmentCall) {
+            throw new NonTranspilableModelException($model);
+        }
+
+        return $this->getAssertionCall(
+            (string) $model->getComparison(),
+            $examinedValueAssignmentCall,
+            $expectedValueAssignmentCall
+        );
+    }
+}
