+++ src/Assertion/IsComparisonTranspiler.php
@@ -3,30 +3,14 @@
 namespace webignition\BasilTranspiler\Assertion;
 
 use webignition\BasilModel\Assertion\AssertionComparisons;
-use webignition\BasilModel\Assertion\AssertionInterface;
 use webignition\BasilTranspiler\CallFactory\AssertionCallFactory;
 use webignition\BasilTranspiler\CallFactory\VariableAssignmentCallFactory;
+use webignition\BasilTranspiler\Model\Call\VariableAssignmentCall;
 use webignition\BasilTranspiler\Model\TranspilationResultInterface;
-use webignition\BasilTranspiler\Model\VariablePlaceholder;
-use webignition\BasilTranspiler\NonTranspilableModelException;
 use webignition\BasilTranspiler\TranspilerInterface;
 
-class IsComparisonTranspiler implements TranspilerInterface
+class IsComparisonTranspiler extends AbstractTwoValueComparisonTranspiler implements TranspilerInterface
 {
-    private $assertionCallFactory;
-    private $variableAssignmentCallFactory;
-    private $assertableValueExaminer;
-
-    public function __construct(
-        AssertionCallFactory $assertionCallFactory,
-        VariableAssignmentCallFactory $variableAssignmentCallFactory,
-        AssertableValueExaminer $assertableValueExaminer
-    ) {
-        $this->assertionCallFactory = $assertionCallFactory;
-        $this->variableAssignmentCallFactory = $variableAssignmentCallFactory;
-        $this->assertableValueExaminer = $assertableValueExaminer;
-    }
-
     public static function createTranspiler(): IsComparisonTranspiler
     {
         return new IsComparisonTranspiler(
@@ -36,74 +20,21 @@ class IsComparisonTranspiler implements TranspilerInterface
         );
     }
 
-    public function handles(object $model): bool
+    protected function getHandledComparisons(): array
     {
-        if (!$model instanceof AssertionInterface) {
-            return false;
-        }
-
-        return in_array($model->getComparison(), [
+        return [
             AssertionComparisons::IS,
             AssertionComparisons::IS_NOT,
-        ]);
+        ];
     }
 
-    /**
-     * @param object $model
-     *
-     * @return TranspilationResultInterface
-     *
-     * @throws NonTranspilableModelException
-     */
-    public function transpile(object $model): TranspilationResultInterface
-    {
-        if (!$model instanceof AssertionInterface) {
-            throw new NonTranspilableModelException($model);
-        }
-
-        $isHandledComparison = in_array($model->getComparison(), [
-            AssertionComparisons::IS,
-            AssertionComparisons::IS_NOT,
-        ]);
-
-        if (false === $isHandledComparison) {
-            throw new NonTranspilableModelException($model);
-        }
-
-        $examinedValue = $model->getExaminedValue();
-        if (null === $examinedValue || !$this->assertableValueExaminer->isAssertableExaminedValue($examinedValue)) {
-            throw new NonTranspilableModelException($model);
-        }
-
-        $expectedValue = $model->getExpectedValue();
-        if (null === $expectedValue || !$this->assertableValueExaminer->isAssertableExpectedValue($expectedValue)) {
-            throw new NonTranspilableModelException($model);
-        }
-
-        $examinedValuePlaceholder = new VariablePlaceholder('EXAMINED_VALUE');
-        $examinedValueAssignmentCall = $this->variableAssignmentCallFactory->createValueVariableAssignmentCall(
-            $examinedValue,
-            $examinedValuePlaceholder
-        );
-
-        $expectedValuePlaceholder = new VariablePlaceholder('EXPECTED_VALUE');
-        $expectedValueAssignmentCall = $this->variableAssignmentCallFactory->createValueVariableAssignmentCall(
-            $expectedValue,
-            $expectedValuePlaceholder
-        );
-
-        if (null === $expectedValueAssignmentCall || null === $examinedValueAssignmentCall) {
-            throw new NonTranspilableModelException($model);
-        }
-
-        return $model->getComparison() === AssertionComparisons::IS
-            ? $this->assertionCallFactory->createValuesAreEqualAssertionCall(
-                $examinedValueAssignmentCall,
-                $expectedValueAssignmentCall
-            )
-            : $this->assertionCallFactory->createValuesAreNotEqualAssertionCall(
-                $examinedValueAssignmentCall,
-                $expectedValueAssignmentCall
-            );
+    protected function getAssertionCall(
+        string $comparison,
+        VariableAssignmentCall $examinedValue,
+        VariableAssignmentCall $expectedValue
+    ): TranspilationResultInterface {
+        return $comparison === AssertionComparisons::IS
+            ? $this->assertionCallFactory->createValuesAreEqualAssertionCall($examinedValue, $expectedValue)
+            : $this->assertionCallFactory->createValuesAreNotEqualAssertionCall($examinedValue, $expectedValue);
     }
 }
